+++ controllers/controllers_suite_test.go
@@ -33,6 +33,7 @@ import (
 	"golang.org/x/tools/go/packages"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/kubectl/pkg/scheme"
 	capa "sigs.k8s.io/cluster-api-provider-aws/v2/api/v1beta2"
 	eks "sigs.k8s.io/cluster-api-provider-aws/v2/controlplane/eks/api/v1beta2"
@@ -42,6 +43,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/envtest/komega"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	karpv1 "sigs.k8s.io/karpenter/pkg/apis/v1"
 
 	// +kubebuilder:scaffold:imports
 
@@ -110,6 +112,10 @@ var _ = BeforeSuite(func() {
 
 	err = eks.AddToScheme(scheme.Scheme)
 	Expect(err).NotTo(HaveOccurred())
+
+	karpv1GV := schema.GroupVersion{Group: "karpenter.sh", Version: "v1"}
+	scheme.Scheme.AddKnownTypes(karpv1GV, &karpv1.NodePool{}, &karpv1.NodePoolList{}, &karpv1.NodeClaim{}, &karpv1.NodeClaimList{})
+	metav1.AddToGroupVersion(scheme.Scheme, karpv1GV)
 
 	k8sClient, err = client.New(cfg, client.Options{Scheme: scheme.Scheme})
 	Expect(err).NotTo(HaveOccurred())