+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use the Karpenter `NodePool` CRD instead of `unstructured.Unstructured` to avoid constant API Update requests.
+
 ## [0.26.1] - 2026-03-16
 
 ### Fixed